Abstract logic of fixing `do` notation into decorator;
To fix issue with state full generators we introduced `IGNORE_VALUE` and updated logic of `foldMap` to count for this value.
Here I have abstracted that logic into a decorator and used this with foldMap so it now looks as before.
It will be useful with other functions which traverse Free structure and need to ignore `IGNORE_VALUE`.